# Faster (still slow) fallback matrix multiplication