@christiangnrd 2025-04-19
Based on https://github.com/JuliaGPU/KernelAbstractions.jl/pull/590, maybe this should wait until GPUArrays supports KA v