@christiangnrd 2026-08-25
OpenCL POCL before and after:

| eltype | size | main | PR | speedup |
|--------|------|-----:|---:|--------:|
| Float16 | 512³  | 23 | 45 | 2.0× |
| Float16 | 1024³ | 22 | 47 | 2.1× |
| Float16 | 2048³ | 13 | 45 | 3.5× |
| Float16 | 4096³ | 2  | 45 | 22.5× |
| Int32   | 512³  | 24 | 60 | 2.5× |
| Int32   | 1024³ | 13 | 60 | 4.6× |
| Int32   | 2048³ | 4  | 59 | 14.8× |
| Int32   | 4096³ | 2  | 37 | 18.5× |
| Float32 | 512³  | 21 | 44 | 2.1× |
| Float32 | 1024³ | 13 | 46 | 3.5× |
| Float32 | 2048³ | 4  | 44 | 11.0× |
| Float32 | 4096³ | 2  | 31 | 15.5× |