@christiangnrd 2026-08-25
I updated this to include the overflow and other various improvements recently applied to the fallback matmul, and reimplemented on top of KernelAbstractions instead of KernelInterface so we can merge sooner rather than later.